Rotated glyphs (#559)
* asset: Store ftx glyph index as a u16

* ui: Pack Ui glyph atlasIndex in 16 bit

* rend: Split fragment math from normal math utils

* ui: Support angle in glyph rendering

* ui: Fix skewing rotated glyphs

* ui: Add angle to draw glyph commands

* ui: Support outputting rotated glyphs

* ui: Verify that rotated glyphs cannot be tracked

* ui: Add angle to ui inspector

* ui: Add various assets for rotated glyphs

* ui: Add ui_line util

* ui: Add circle util

* rend: Draw fog when fog-debug is enabled

+++ libs/asset/include/asset_ftx.h
@@ -17,7 +17,7 @@
 typedef struct {
   Unicode cp;
   u8      variation;
-  u32     glyphIndex; // sentinel_u32 when character has no glyph (for example a space).
+  u16     glyphIndex; // sentinel_u16 when character has no glyph (for example a space).
   f32     size;
   f32     offsetX, offsetY;
   f32     advance;
